Refactor PWAEvent to context
